chatbot/backend/requirements.txt
langchain
langchain-groq
langchain-core
langchain-community
langchain-cohere
langchain-chroma
fastapi
pydantic
uvicorn
sse_starlette
python-dotenv